рефакторинг на основе замечаний
