+++ src/const/filter-const.js
@@ -1,0 +1,21 @@
+import { isActualPoint, isExpiredPoint, isFuturePoint } from '../utils/points-utils';
+
+
+const FiltersPoint = {
+  EVERYTHING: 'everything',
+  FUTURE: 'future',
+  PRESENT: 'present',
+  PAST: 'past'
+};
+
+const filter = {
+  [FiltersPoint.EVERYTHING]: (points) => points,
+  [FiltersPoint.FUTURE]: (points) => points.filter((point) => isFuturePoint(point)),
+  [FiltersPoint.PRESENT]: (points) => points.filter((point) => isActualPoint(point)),
+  [FiltersPoint.PAST]: (points) => points.filter((point) => isExpiredPoint(point))
+};
+
+export {
+  FiltersPoint,
+  filter
+};
