+++ src/main.js
@@ -6,13 +6,13 @@ import { render } from './framework/render.js';
 import PointsModel from './model/point-model.js';
 import OffersModel from './model/offers-model.js';
 import DestinationsModel from './model/destinations-model.js';
-import { generationFilters } from './mocks/filters-mock.js';
+import { generateFilters } from './mocks/filters-mock.js';
 
 const pointsModel = new PointsModel();
 const offersModel = new OffersModel();
 const destinationsModel = new DestinationsModel();
 
-const filters = generationFilters(pointsModel.points);
+const filters = generateFilters(pointsModel.points);
 
 const boardContainer = document.querySelector('.trip-events');
 const headerContainer = document.querySelector('.trip-main');
