+++ src/utils/points-utils.js
@@ -1,18 +1,9 @@
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
-import { EDIT_POINT_DATE_FORMAT, POINT_DATE_FORMAT } from './const';
+import { EDIT_POINT_DATE_FORMAT, POINT_DATE_FORMAT } from '../const/points-const';
+import { getRandomInt } from './utils';
 dayjs.extend(duration);
 
-const getRandomInt = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const findById = (arr, id) => arr.find((element) => element.id === id);
-
-const getRandomArrayElement = (items) => items[Math.floor(Math.random() * items.length)];
 
 const totalOffersPrice = (offers) => offers.reduce((accumulator, currentValue) => accumulator + currentValue.price, 0);
 
@@ -22,7 +13,6 @@ const formatPointDate = (pointDate) => pointDate.format(POINT_DATE_FORMAT);
 
 const getDurationEvent = (dateFrom, dateTo) => dayjs.duration(dateTo.diff(dateFrom));
 
-const capitalizeFirstLetter = (str) => str.charAt(0).toUpperCase() + str.slice(1);
 
 const getOffersByType = (offers, type) => {
   const offer = offers.filter((offerItem) => offerItem.type === type);
@@ -50,14 +40,10 @@ const isActualPoint = (point) => point.dateTo && (dayjs().isSame(dayjs(point.dat
 
 
 export {
-  getRandomInt,
-  findById,
-  getRandomArrayElement,
   totalOffersPrice,
   formatEditPointDate,
   formatPointDate,
   getDurationEvent,
-  capitalizeFirstLetter,
   getOffersByType,
   getRandomDate,
   isFuturePoint,
